More cleanup and reorganization.
